后端一些修改

+++ src/main/java/top/sshh/bililiverecoder/service/impl/LiveMsgService.java
@@ -51,22 +51,31 @@ public class LiveMsgService {
     RecordHistoryPartRepository partRepository;
 
     public int sendMsg(BiliBiliUser user, LiveMsg liveMsg) {
-        BiliDmResponse response = BiliApi.sendVideoDm(user, liveMsg);
-        int code = response.getCode();
-        if (code != 0) {
-            log.error("{}发送弹幕错误，code==>{}", user.getUname(), code);
-            if (code == 36701 || code == 36702 || code == 36714) {
-                liveMsgRepository.delete(liveMsg);
+        try {
+            BiliDmResponse response = BiliApi.sendVideoDm(user, liveMsg);
+            if (response == null) {
+                log.error("{}发送弹幕失败，响应为空", user.getUname());
+                return -1;
             }
-            if(code == 36704){
-                String bvid = liveMsg.getBvid();
-                this.syncVideoState(bvid);
-                return code;
+            int code = response.getCode();
+            if (code != 0) {
+                log.error("{}发送弹幕错误，code==>{} msg==>{}", user.getUname(), code, response.getMessage());
+                if (code == 36701 || code == 36702 || code == 36714) {
+                    liveMsgRepository.delete(liveMsg);
+                }
+                if(code == 36704){
+                    String bvid = liveMsg.getBvid();
+                    this.syncVideoState(bvid);
+                    return code;
+                }
             }
+            liveMsg.setCode(code);
+            liveMsgRepository.save(liveMsg);
+            return code;
+        } catch (Exception e) {
+            log.error("{}发送弹幕异常: {}", user.getUname(), e.getMessage());
+            return -2;
         }
-        liveMsg.setCode(code);
-        liveMsgRepository.save(liveMsg);
-        return code;
     }
 
     public static boolean checkUtf8Size(String testStr) {
@@ -141,7 +150,7 @@ public class LiveMsgService {
                     msg.setFontsize(64);
                     msg.setColor(16776960);
                     StringBuilder builder = new StringBuilder();
-                    builder.append(userName).append("发送了").append(price).append("元留言：").append(text);
+                    builder.append("SC [").append(price).append("] ").append(userName).append(": ").append(text);
                     if (builder.length() > 100) {
                         text = builder.substring(0, 99);
                     } else {
@@ -169,16 +178,16 @@ public class LiveMsgService {
                     msg.setPool(1);
                     msg.setColor(16776960);
                     StringBuilder builder = new StringBuilder();
-                    builder.append(userName).append("开通了");
+                    builder.append("⚓").append(userName).append(": 开通了");
                     if (Integer.parseInt(count) > 1) {
                         builder.append(count).append("个月");
                     }
                     if ("1".equals(level)) {
                         msg.setFontsize(64);
-                        builder.append("19998/月的总督");
+                        builder.append("总督");
                     } else if ("2".equals(level)) {
                         msg.setFontsize(64);
-                        builder.append("1998/月的提督");
+                        builder.append("提督");
                     } else if ("3".equals(level)) {
                         msg.setFontsize(64);
                         builder.append("舰长");
